fix(Snakemake): create plot_dir via Snakemake
Since it's not possible to commit an empty directory, the project
structure is incomplete on cloning, and ggplot2 *will* complain. By
forcing one of the earlier run rules to require the plot directory, we
can make explicit the fact that these files are being output, as well as
automate the creating of the directory.

Ultimately, it would probably be wise to always reference the plot_dir
variable in any file that produces plots, because currently this depends
on this rule running first. But it's a start.