dev: use renv for R dependencies
I decided to use renv for R dependencies. Unfortunately, it doesn't play
very nicely with Snakemake. Snakemake runs R files with the --vanilla
flag, with (as best I can tell) no easy way to change that. As such, the
.Rprofile won't get sourced, which is how {renv} typically gets
activated. To counteract this, I just added source("renv/activate.R") at
the beginning of each script. It's a bit clunky, but I suppose it works.